Text Text Buttons
Added a component that adds buttons to inspector of the instructions text object that displays questions for the test participant during the pauses between test conditions.
These buttons allow for an easy 1-click solution in the editor, removing the need for a test technician to copy and paste questions from another application.

ToDo, make the text automatically swap back to the default "wait" text when the test is unpaused