Mole List Generator
Made a scene that can generate a list of mole inputs for the Wampat file format, features customizable size, automatic duplicate protection and semi-insurance of balanced distribution.
Made moles reset when a new modifier is applied so depleted moles are returned to normal.
Made game trigger STOP automatically when the test is done and apply a fade to signal the user.